Move shared classes that shouldn't be exposed to the common module
