+++ common/src/main/java/com/ably/tracking/common/Ably.kt
@@ -1,13 +1,9 @@
 package com.ably.tracking.common
 
 import com.ably.tracking.ConnectionException
-import com.ably.tracking.ConnectionStateChange
 import com.ably.tracking.EnhancedLocationUpdate
 import com.ably.tracking.LocationUpdate
-import com.ably.tracking.clientOptions
 import com.ably.tracking.connection.ConnectionConfiguration
-import com.ably.tracking.toTracking
-import com.ably.tracking.toTrackingException
 import com.google.gson.Gson
 import io.ably.lib.realtime.AblyRealtime
 import io.ably.lib.realtime.Channel
